simplifyConjunctions: move to condition simplifier
