+++ kore/src/Kore/Step/Step.hs
@@ -72,7 +72,6 @@ import Kore.Rewriting.UnifyingRule
 import qualified Kore.Step.Result as Result
 import qualified Kore.Step.Result as Results
 import qualified Kore.Step.Result as Step
-import qualified Kore.Step.Simplification.Pattern as Pattern
 import Kore.Step.Simplification.Simplify
     ( MonadSimplify
     )
@@ -300,4 +299,3 @@ applyRemainder initial remainder = do
     Simplifier.simplifyCondition
         SideCondition.topTODO
         (Pattern.andCondition initial partial)
-        >>= return . Pattern.simplifyConjunctions
